Fix protracted audio play issue
We found out that if we use the "s1 play -start <start> -end <stop>..."
regularly the audio clip will not stop at <stop> but rather a second
after that. To get around this pet peeve, we destroy the sound object
and recreate it during the audio gap in between each sentence.